requirements.txt
streamlit
PyPDF2
google.generativeai

python-dotenv